+++ apps/vscode-extension/src/webview/WebviewHtml.ts
@@ -5,7 +5,7 @@ export const WebviewHtml = `<!doctype html>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Demo Time Webview</title>
-    <script type="module" crossorigin src="./assets/index-tM0esR6J.js"></script>
+    <script type="module" crossorigin src="./assets/index-CcthKTPs.js"></script>
     <link rel="stylesheet" crossorigin href="./assets/index-DLGA4Yij.css">
   </head>
   <body>