+++ docs/astro.config.mjs
@@ -217,4 +217,9 @@ export default defineConfig({
     }), 
     react()
   ],
+  vite: {
+		ssr: {
+			noExternal: ['zod'],
+		},
+	}
 });